Get image id from image name to avoid hardcoding it as it depends on the ami

+++ deployment/terminate_instances.py
@@ -1,20 +1,16 @@
 import boto3
 import config as cfg
 from tools import is_smack_instance
+from get_instances import get_instances
 
 ec2 = boto3.client('ec2')
 
 
 def terminate_instances():
-    to_terminate = []
-    resp = ec2.describe_instances()
-    reservations = resp['Reservations']
-    for reservation in reservations:
-        for instance in reservation['Instances']:
-            if is_smack_instance(instance):
-                to_terminate.append(instance['InstanceId'])
-
-    ec2.terminate_instances(InstanceIds=to_terminate)
+    to_terminate = [instance['InstanceId'] for instance in get_instances()]
+    if to_terminate:
+        ec2.terminate_instances(InstanceIds=to_terminate)
+    print("{0} instances terminated.".format(len(to_terminate)))
 
 
 if __name__ == '__main__':
